refactor: move pseudo to internal
