+++ core/state/state.libevm.go
@@ -19,7 +19,7 @@ package state
 import (
 	"github.com/ava-labs/libevm/common"
 	"github.com/ava-labs/libevm/core/types"
-	"github.com/ava-labs/libevm/libevm/pseudo"
+	"github.com/ava-labs/libevm/internal/libevm/pseudo"
 )
 
 // GetExtra returns the extra payload from the [types.StateAccount] associated
