+++ core/types/block.libevm.go
@@ -20,7 +20,7 @@ import (
 	"encoding/json"
 	"io"
 
-	"github.com/ava-labs/libevm/libevm/pseudo"
+	"github.com/ava-labs/libevm/internal/libevm/pseudo"
 	"github.com/ava-labs/libevm/rlp"
 )
